requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
bitstring==3.1.9
certifi==2021.10.8
cffi==1.14.6
google-cloud-texttospeech==2.11.1
nltk==3.7
charset-normalizer==2.0.12
colorama==0.4.4
cryptography==3.4.7
cycler==0.11.0
dj-database-url==0.5.0
Django==4.0.2
django-countries==7.3.2
django-heroku==0.3.1
ecdsa==0.17.0
empy==3.3.4
esptool==3.1
et-xmlfile==1.1.0
fonttools==4.31.2
gunicorn==20.1.0
heroku3==5.1.4
idna==3.3
joblib==1.1.0
kiwisolver==1.4.0
lxml==4.6.2
matplotlib==3.5.1
numpy==1.20.1
openai==0.16.0
openpyxl==3.0.9
django-extensions==3.1.5
pydot==1.4.2
packaging==21.3
pandas==1.4.1
pandas-stubs==1.2.0.53
passlib==1.7.4
Pillow==9.0.1
psycopg2==2.9.3
pycparser==2.20
pyparsing==3.0.7
pyserial==3.5
python-dateutil==2.8.2
python-decouple==3.6
pytz==2021.3
reedsolo==1.5.4
requests==2.27.1
scikit-learn==1.0.2
scipy==1.7.3
six==1.16.0
sklearn==0.0
sqlparse==0.4.2
threadpoolctl==3.0.0
tqdm==4.63.0
typing_extensions==4.1.1
tzdata==2021.5
urllib3==1.26.9
whitenoise==6.0.0